+++ packages/docs/site/rollup.config.js
@@ -1,13 +1,27 @@
+import path from 'node:path';
+import {fileURLToPath} from 'node:url';
+
 import json from '@rollup/plugin-json';
 import resolve from '@rollup/plugin-node-resolve';
 import replace from '@rollup/plugin-replace';
 import commonjs from '@rollup/plugin-commonjs';
+import alias from 'rollup-plugin-alias';
 import svelte from 'rollup-plugin-svelte';
 import babel from 'rollup-plugin-babel';
 import {terser} from 'rollup-plugin-terser';
 import yaml from '@rollup/plugin-yaml';
 
 import config from 'sapper/config/rollup.js';
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
+
+const appRoot = path.join(__dirname, 'src/lib');
+const aliasConfig = alias({
+	resolve: ['.js', 'svelte'],
+	entries: [
+		{find: '$lib', replacement: appRoot}
+	]
+});
 
 const mode = process.env.NODE_ENV;
 const dev = mode === 'development';
@@ -43,6 +57,7 @@ export default {
 			commonjs(),
 			json(),
 			yaml(),
+			aliasConfig,
 
 			legacy && babel({
 				extensions: ['.js', '.mjs', '.html', '.svelte'],
@@ -93,6 +108,7 @@ export default {
 			commonjs(),
 			json(),
 			yaml(),
+			aliasConfig,
 		],
 		preserveEntrySignatures,
 	},
@@ -110,6 +126,7 @@ export default {
 			commonjs(),
 			json(),
 			yaml(),
+			aliasConfig,
 			!dev && terser()
 		],
 		preserveEntrySignatures,