+++ packages/docs/site/src/bin/makeEntries.js
@@ -1,16 +1,32 @@
-#!/usr/bin/env node -r svelte/register
+#!/usr/bin/env node
 
-import {concatValues, joinWithBlank, makePrefixed} from '@svizzle/utils';
+import {joinWithBlank, makePrefixed} from '@svizzle/utils';
 import * as _ from 'lamb';
 
-import * as examples from '../routes/components/_examples/index.js';
+import {default as barchart} from '../routes/components/_examples/barchart.js';
+import {default as choropleth} from '../routes/components/_examples/choropleth.js';
+import {default as histogram} from '../routes/components/_examples/histogram.js';
+import {default as legend} from '../routes/components/_examples/legend.js';
+
+const slugs = [
+	...barchart,
+	...choropleth,
+	...histogram,
+	...legend,
+]
+.map(_.getKey('slug'))
+.concat([
+	'ui-Icon',
+	'ui-Switch',
+	'ui-Link',
+	'ui-LinkButton',
+	'ui-LoadingView'
+])
 
 const makeEntries = _.pipe([
-	concatValues,
-	_.pluck('slug'),
 	_.mapWith(makePrefixed('components/')),
 	joinWithBlank
 ]);
 
-const entries = `'components ${makeEntries(examples)}'`;
+const entries = `'components ${makeEntries(slugs)}'`;
 console.log(entries);