+++ CHANGELOG.md
@@ -125,12 +125,14 @@
 			- set `engines.node` to `>=17.5.0`
 			- moved some dev deps in `dependencies`
 		- deps:
-			- upgraded `chance`, `d3-array`, `d3-color`, `d3-dsv`, `d3-geo`, `d3-scale`, `eslint`, `eslint-plugin-svelte3`, `svelte-json-tree`
+			- upgraded all dependencies
 			- removed `esm`, `eslint-plugin-import`
 	- imports:
 		- using file extensions
 		- use the `node:` protocol
 - now internally importing directly from svizzle modules index
+- moved `src/node_modules` to `src/lib` and added `rollup-plugin-alias`
+- updated scripts
 
 ## `@svizzle/atlas` v0.8.0 (next)